Split back fetch more loader
